+++ lib/td_dd/grants/grant_request_group.ex
@@ -12,6 +12,7 @@ defmodule TdDd.Grants.GrantRequestGroup do
     field(:type, :string)
     field(:user_id, :integer)
     field(:user, :map, virtual: true)
+    field(:created_by_id, :integer)
 
     has_many(:requests, GrantRequest, foreign_key: :group_id)
     belongs_to(:modification_grant, Grant)
@@ -27,7 +28,7 @@ defmodule TdDd.Grants.GrantRequestGroup do
 
   def changeset(%__MODULE__{} = struct, params, _) do
     struct
-    |> cast(params, [:type])
+    |> cast(params, [:type, :user_id, :created_by_id])
     |> cast_requests()
   end