+++ lib/td_dd_web/resolvers/domains.ex
@@ -107,6 +107,9 @@ defmodule TdDdWeb.Resolvers.Domains do
   defp permitted_domain_ids(%{role: "user", jti: jti}, "shareConcept"),
     do: Permissions.permitted_domain_ids(jti, :view_domain)
 
+  defp permitted_domain_ids(%{role: "user", jti: jti}, "createForeignGrantRequest"),
+    do: Permissions.permitted_domain_ids(jti, :create_foreign_grant_request)
+
   defp permitted_domain_ids(%{role: "user", jti: jti}, action_or_permission),
     do: Permissions.permitted_domain_ids(jti, Macro.underscore(action_or_permission))